requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
nextcord==2.2.0
nextcord-ext-activities==2022.6.11
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.2
typing_extensions==4.3.0
urllib3==1.26.14
yarl==1.8.1
